+++ lib/terminal_app/ffmpeg/ffmpeg_audio.dart
@@ -1,6 +1,8 @@
+import 'dart:convert';
 import 'dart:io';
 
 import 'package:than_pkg/terminal_app/ffmpeg/ffmpeg_process.dart';
+import 'package:than_pkg/terminal_app/ffmpeg/audio_metadata.dart';
 
 class FFmpegAudio {
   static final FFmpegAudio instance = FFmpegAudio._();
@@ -18,6 +20,76 @@ class FFmpegAudio {
     }
     final arguments = ['-i', inputPath, outputPath];
     await ffmpegStart(arguments: arguments, onStdErrorOut: onStdout);
+  }
+
+  ///
+  ///  ## Audio Metadata
+  ///
+  Future<AudioMetadata> getAudioMetadata(String audioPath) async {
+    //ffprobe -v quiet -show_entries format_tags -of json
+    final result = await ffprobeRun(
+      arguments: [
+        '-v',
+        'quiet',
+        '-show_entries',
+        'format_tags',
+        '-of',
+        'json',
+        audioPath,
+      ],
+    );
+
+    if (result.exitCode != 0) {
+      throw Exception('ffprobe failed: ${result.stderr}');
+    }
+
+    final jsonData = jsonDecode(result.stdout);
+
+    // format.tags ထဲက metadata ပဲ return လုပ်တာ
+    final tags = jsonData['format']?['tags'] as Map<String, dynamic>? ?? {};
+
+    return AudioMetadata.fromMap(tags);
+  }
+
+  ///
+  /// ## Audio Metadata Edit
+  ///
+  Future<void> editAudioMetadata({
+    required String inputPath,
+    required String outputPath,
+    required AudioMetadata metadata,
+    bool isOverride = true,
+  }) async {
+    final outFile = File(outputPath);
+    if (outFile.existsSync() && isOverride) {
+      await outFile.delete();
+    }
+
+    final args = <String>['-i', inputPath];
+
+    if (metadata.title != null) {
+      args.addAll(['-metadata', 'title=${metadata.title}']);
+    }
+    if (metadata.artist != null) {
+      args.addAll(['-metadata', 'artist=${metadata.artist}']);
+    }
+    if (metadata.album != null) {
+      args.addAll(['-metadata', 'album=${metadata.album}']);
+    }
+    if (metadata.genre != null) {
+      args.addAll(['-metadata', 'genre=${metadata.genre}']);
+    }
+    if (metadata.date != null) {
+      args.addAll(['-metadata', 'date=${metadata.date}']);
+    }
+    if (metadata.comment != null) {
+      args.addAll(['-metadata', 'comment=${metadata.comment}']);
+    }
+
+    // args.addAll(['-c', 'copy', outputPath]);
+    args.addAll(['-id3v2_version', '3', outputPath]);
+
+    await ffmpegRun(arguments: args);
   }
 
   // Future<void> extractAudio() async {}