+++ lib/utils/f_path.dart
@@ -1,7 +1,5 @@
 import 'dart:io';
 
-class FPath {
-  static String join(String path, String name) {
-    return '$path${Platform.pathSeparator}$name';
-  }
+String pathJoin(String path, String name) {
+  return '$path${Platform.pathSeparator}$name';
 }