+++ lib/utils/index.dart
@@ -1,1 +1,2 @@
 export './app_util.dart';
+export './f_path.dart';